libft

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: iounejja <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/16 19:44:05 by iounejja          #+#    #+#             */
-/*   Updated: 2019/10/31 14:33:06 by iounejja         ###   ########.fr       */
+/*   Updated: 2019/11/01 12:35:10 by iounejja         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while ((s1[i] || s2[i]) && i < n)
